+++ tests/misc/programs/misc/src/lib.rs
@@ -189,4 +189,12 @@ pub mod misc {
         ctx.accounts.data.data = data;
         Ok(())
     }
+
+    pub fn test_multidimensional_array(
+        ctx: Context<TestMultidimensionalArray>,
+        data: [[u8; 10]; 10],
+    ) -> ProgramResult {
+        ctx.accounts.data.data = data;
+        Ok(())
+    }
 }